requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.43.3
beautifulsoup4==4.12.3
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
EbookLib==0.18
Eel==0.16.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.7
lxml==5.2.2
packaging==24.0
pdf2image==1.17.0
pefile==2023.2.7
pillow==10.3.0
pycparser==2.22
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.2
PyPDF2==3.0.1
pywin32-ctypes==0.2.2
reportlab==4.2.0
requests==2.32.2
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.4.post2
